Revising directories structure
