+++ src/app.js
@@ -6,26 +6,26 @@ import {
 } from 'react-router-dom';
 
 import './app.css';
-import appSettings from './app/AppSettings';
-import Header from './app/Header';
-import Footer from './app/Footer';
-import Section from './app/Section';
-import AboutUs from './app/AboutUs';
-import FAQ from './app/FAQ';
-import Rates from './app/Rates';
-import LogIn from './app/LogIn';
-import SignUp from './app/SignUp';
+import appSettings from './app/app-settings';
+import Header from './app/header';
+import Footer from './app/footer';
+import Section from './app/section';
+import AboutUs from './app/about-us';
+import FAQ from './app/faq';
+import Rates from './app/rates';
+import LogIn from './app/log-in';
+import SignUp from './app/sign-up';
 
 const App = () => (
   <Router basename={appSettings.base_name}>
     <div className="app">
       <Header />
       <Route exact path="/" component={Section} />
-      <Route path="/AboutUs" component={AboutUs} />
-      <Route path="/FAQ" component={FAQ} />
-      <Route path="/Rates" component={Rates} />
-      <Route path="/LogIn" component={LogIn} />
-      <Route path="/SignUp" component={SignUp} />
+      <Route path="/about-us" component={AboutUs} />
+      <Route path="/faq" component={FAQ} />
+      <Route path="/rates" component={Rates} />
+      <Route path="/log-in" component={LogIn} />
+      <Route path="/sign-up" component={SignUp} />
       <Footer />
     </div>
   </Router>
