+++ src/app/faq/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import '../styles/stationery.css';
-import StationeryWIP from './StationeryWIP';
+import '../../styles/stationery.css';
+import StationeryWIP from '../stationery-wip';
 
-const SignUp = function () {
+const FAQ = function () {
   return (
     <div className="stationery">
       <div className="wall"></div>
 
       <div className="wrapper papper">
-        <h1 className="hidden">SignUp</h1>
+        <h1 className="hidden">FAQ</h1>
         <StationeryWIP />
       </div>
     </div>
   );
 };
 
-export default SignUp;
+export default FAQ;
