+++ src/app/footer/styles.scss
