+++ src/app/log-in/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import '../styles/stationery.css';
-import StationeryWIP from './StationeryWIP';
+import '../../styles/stationery.css';
+import StationeryWIP from '../stationery-wip';
 
-const Rates = function () {
+const LogIn = function () {
   return (
     <div className="stationery">
       <div className="wall"></div>
 
       <div className="wrapper papper">
-        <h1 className="hidden">Rates</h1>
+        <h1 className="hidden">Log In</h1>
         <StationeryWIP />
       </div>
     </div>
   );
 };
 
-export default Rates;
+export default LogIn;
