+++ src/app/rate-details/styles.scss
