+++ src/app/select-rate/index.js
