+++ src/app/sign-up/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import '../styles/stationery.css';
-import StationeryWIP from './StationeryWIP';
+import '../../styles/stationery.css';
+import StationeryWIP from '../stationery-wip';
 
-const LogIn = function () {
+const SignUp = function () {
   return (
     <div className="stationery">
       <div className="wall"></div>
 
       <div className="wrapper papper">
-        <h1 className="hidden">Log In</h1>
+        <h1 className="hidden">SignUp</h1>
         <StationeryWIP />
       </div>
     </div>
   );
 };
 
-export default LogIn;
+export default SignUp;
